Add cstructs.
As I'm doing this, I realize it feels nontrivial to add cstructs to a Makefile;
in general, it seems nontrivial to add C project dependencies without a package
manager. I wonder if there's a nice way to handle the Makefile part of the
import, as for many of my libraries, the only non-Makefile importing work is
simply copying files.