Switch to an array of lines as the main buffer.
I realized this is a particularly nice way to hold the data as the final-newline
question can be resolved by simply considering the array the "\n"-split of the
file. So the "\n"-join gives the file, and it ends in a newline iff the final
line in our array is the empty string.