Fix the append command.
I just had an off-by-one error.

I also noticed that ed2's = command deviates a little from ed's behavior in that
a command line like "2,5=" will report 5 in ed but *not* change the current
line. However, in ed2, it *does* change the current line. I've decided not to
fix this as, so far, ed2's behavior seems more internally consistent to me.